DIrectly use Native libraries instead of extra GramineNative Library
