+++ conclave-host/src/main/kotlin/com/r3/conclave/host/internal/attestation/EnclaveQuoteServiceDCAP.kt
@@ -9,7 +9,7 @@ class EnclaveQuoteServiceDCAP : EnclaveQuoteService() {
 
     override fun initializeQuote(): Cursor<SgxTargetInfo, ByteBuffer> {
         val targetInfo = Cursor.allocate(SgxTargetInfo)
-        Native.initQuoteDCAP(NativeLoader.libsPath.toString(), targetInfo.buffer.array(), true)
+        Native.initQuoteDCAP(NativeLoader.libsPath.toString(), targetInfo.buffer.array())
         return targetInfo
     }
