+++ cpp/jvm-host/include/dcap.h
@@ -29,13 +29,13 @@ namespace r3::conclave::dcap {
 
         sgx_ql_qve_collateral_t* collateral;
 
-        SGX_QL_DECL(sgx_ql_set_path, sgx_ql_path_type_t, const char *);
-        SGX_QL_DECL(sgx_qe_set_enclave_load_policy, sgx_ql_request_policy_t);
+        SGX_QL_DECL(sgx_ql_set_path,sgx_ql_path_type_t,const char *);
+        SGX_QL_DECL(sgx_qe_set_enclave_load_policy,sgx_ql_request_policy_t);
         SGX_QL_DECL(sgx_qe_cleanup_by_policy);
 
         SGX_QL_DECL(sgx_qe_get_target_info, sgx_target_info_t*);
-        SGX_QL_DECL(sgx_qe_get_quote_size, uint32_t*);
-        SGX_QL_DECL(sgx_qe_get_quote, sgx_report_t *, uint32_t, uint8_t*);
+        SGX_QL_DECL(sgx_qe_get_quote_size,uint32_t*);
+        SGX_QL_DECL(sgx_qe_get_quote,sgx_report_t *,uint32_t,uint8_t*);
 
         SGX_QL_DECL(sgx_ql_get_quote_verification_collateral, const uint8_t*, uint16_t, const char* pck_ca, sgx_ql_qve_collateral_t**);
         SGX_QL_DECL(sgx_ql_free_quote_verification_collateral, sgx_ql_qve_collateral_t *p_quote_collateral);
@@ -43,7 +43,7 @@ namespace r3::conclave::dcap {
     public:
         typedef std::vector<std::string> Errors;
 
-        bool init(const std::string& path, Errors& errors, const bool skipQuotingLibraries);
+        bool init(const std::string& path, Errors& errors);
 
         bool get_target_info(sgx_target_info_t* target_info, quote3_error_t& eval_result);
         bool get_quote_size(uint32_t* p_size, quote3_error_t& eval_result);
