+++ cpp/jvm-host/src/dcap.cpp
@@ -56,9 +56,7 @@ namespace r3::conclave::dcap {
         throw new std::exception(); // fatal, not suppose to happen
     }
 
-    bool QuotingAPI::init(const std::string& path,
-                          QuotingAPI::Errors& errors,
-                          const bool skipQuotingLibraries) {
+    bool QuotingAPI::init(const std::string& path, QuotingAPI::Errors& errors) {
         //  The users can use SGX_AESM_ADDR environment variable to select
         //    whether some functions of the quoting library run "in-process" or "out-of-process.
         //  Conclave uses the "in-process" approach, and the SGX_AESM_ADDR variable
